add types.ts file
