+++ src/types.ts
@@ -1,0 +1,13 @@
+export type DatepickerType = Date | null;
+
+export type CropType = {
+  id: number;
+  name: string;
+  isPlanted: boolean;
+  seededAt: DatepickerType;
+  harvestedAt: DatepickerType;
+};
+
+export type CropsType = {
+  [id: number]: CropType;
+};
